refactor(finance): share monthly-report logic between API and worker (de-dup #479)
PR #479 fixed the setTimeout busy-loop by moving the monthly finance-report
cron to a BullMQ repeatable worker job, but duplicated ~814 lines of finance
business logic into packages/worker/src/services/finance/{period,mobilisation,
summary}.ts and re-implemented the report orchestration inline. That risks the
worker's kpi_snapshot silently drifting from the API/dashboard's.

De-duplicate into one dependency-injected shared module that BOTH the API
(manual report + dashboard) and the worker (cron) call:

- packages/shared/src/finance/reporting/{period,mobilisation,summary,monthly-report}.ts
  — buildFinanceSummary(db, …) + requestMonthlyReport(deps, …) take an injected
  db handle + enqueue closure (same DI pattern as shared/finance/exchange-rate-service
  + signup/resend). Exposed via the dedicated `@givernance/shared/finance/reporting`
  subpath — NOT the root barrel, so no Drizzle-query code leaks into the web
  bundle (ADR-013).
- API period/mobilisation/service/monthly-report become thin shims that pre-bind
  systemDb + the API PLATFORM_REPORTS queue — routes/tests unchanged, manual mode intact.
- Worker platform-report-trigger.ts becomes a thin adapter: flag-gate, inject the
  worker owner-pool db + enqueue closure, call the shared request/backfill.
- Delete the three worker-side duplicate modules.
- Add packages/worker/src/processors/platform-report-trigger.test.ts (flag-off no-op,
  single→request, backfill→backfill) — #479 had no automated processor test.

typecheck + biome + build green; worker/shared/api tests pass.
