+++ packages/shared/package.json
@@ -11,6 +11,7 @@
     "./jobs": "./src/jobs/index.ts",
     "./validators": "./src/validators/index.ts",
     "./constants": "./src/constants/index.ts",
+    "./finance/reporting": "./src/finance/reporting/index.ts",
     "./i18n": "./src/i18n/locales.ts",
     "./contracts/branding": "./src/contracts/branding.ts",
     "./postal-export-mode": "./src/postal-export-mode.ts",
